apps/synthesizer/pyproject.toml
[project]
name = "synthesizer"
version = "0.1.0"
description = "Kasita content processing and knowledge unit generation"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.39.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "python-shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "processors", "generators", "models"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
]

[tool.uv.sources]
python-shared = { path = "../../libs/python-shared", editable = true }

